Created the basic race logic
Basic race logic is that that each racer 'rolls' a d20 and adds their
modifier. The result is added to their total score. Highest score at the
end of the final round is calculated and displayed. In true DnD fashion,
natural 20s and natural 1s are treated differently and a fun message
gets displayed when they occur.

Also changed the input type in the racers-component because it was
treating the input as a string instead of a proper number.