Ties are now calculated and text is ordered
The race-outcome component now pulls its own list of racers from the
service and then orders them based on score. When the 'announcer' lines
are created, they will be read in that order which is more natural.

I also used this to check for ties at the end of the race .

A little house keeping on whitespace too.