Created the race visualizer
Race visualizer is now in. It keeps track of how long the track is and
then uses a hidden <div> to place a visible <div> at the appropriate
percentage down the track.
Because I wanted to ensure that 5 legs of the race are ran, the highest racer
is assumed to be at the end of the current leg, and the remaining racers'
positions are calculated based on its score.

Also a few style flubs were caught by the linter.