Changed the racer icons to be boxes of colors
Each racer now is represented by a set size square instead of an X or O.

Added a few bits of CSS that will keep the racer's names aligned with
their icons when more than 5 racers are added.

Fixed a color issue where was black in cases where a single digit hex
value was selected for one of the RGB values.